% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTierpsyTracks.R
\name{ExtractTierpsyTracks}
\alias{ExtractTierpsyTracks}
\title{Filter and align worms that have been previously marked as good in Tierpsy Tracker}
\usage{
ExtractTierpsyTracks(
  file,
  framerate = 10,
  group = TRUE,
  compare1 = strain,
  compare2 = NULL,
  pixelRatio = 1,
  ...
)
}
\arguments{
\item{file}{this would be a metadata file for a folder to be analyzed or a
single hdf5 file if analyzing a single tracked video}
}
\value{
a time-aligned
}
\description{
Filter and align worms that have been previously marked as good in Tierpsy Tracker
}
\examples{
ExtractTierpsyTracks()

}
