% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForagingAnalysis.R
\name{ForagingAnalysis}
\alias{ForagingAnalysis}
\title{Filter and align worms that have been previously marked as good in Tierpsy Tracker}
\usage{
ForagingAnalysis(
  file,
  framerate = 10,
  min_frames = 50,
  group = TRUE,
  compare1 = strain,
  compare2 = NULL,
  use_only_metadata = FALSE,
  pixelRatio = 1,
  ...
)
}
\arguments{
\item{file}{this would be a metadata file for a folder to be analyzed or a
single hdf5 file if analyzing a single tracked video}
}
\value{
a time-aligned
}
\description{
Filter and align worms that have been previously marked as good in Tierpsy Tracker
}
\examples{
ForagingAnalysis()

}
